Refactored code into modules
There is now a search module for the search pane and an edit module for
editing tiddlers (plus some others).

It's using requirejs for the modular approach, and relies on the
chrjs-store bundling of almond.js for requirejs presence. It's also
using events.js and exemplar.js, 2 micro libraries to handle creating
function exemplars and eventing (inc. DOM events). They can be found at
https://github.com/bengillies/stuff-js